fix: wip rework courtesy contact item and tests
